[finish]destination

+++ app/controllers/public/post_favorites_controller.rb
@@ -1,16 +1,24 @@
 class Public::PostFavoritesController < ApplicationController
-  def create
+ def create
     post = Post.find(params[:post_id])
-    favorite = current_member.favorites.new(post_id: post.id)
-    favorite.save
-    redirect_to post_path(post)
-  end
-  
-  def destroy
+    if current_member.nil?
+      redirect_to login_path, alert: "ログインしてください"
+    else
+      favorite = current_member.favorites.new(post_id: post.id)
+      favorite.save
+      redirect_to post_path(post)
+    end
+ end
+
+ def destroy
     post = Post.find(params[:post_id])
-    favorite = current_member.favorites.find_by(post_id: post.id)
-    favorite.destroy
-    redirect_to post_path(post)
-  end
+    if current_member.nil?
+      redirect_to login_path, alert: "ログインしてください"
+    else
+      favorite = current_member.favorites.find_by(post_id: post.id)
+      favorite.destroy
+      redirect_to post_path(post)
+    end
+ end
 
 end
